requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
python-multipart==0.0.20
alembic==1.16.5
pymupdf==1.26.4
python-docx==1.2.0
openai==1.106.1
anthropic==0.68.0
ollama>=0.4.7
pytest==8.3.3
openpyxl==3.1.5
python-pptx==1.0.2
pyhwp==0.1b15
pyyaml>=6.0.2
httpx>=0.27.0
